backendv2/app/dependencies/auth.py
"""
Authentication Dependencies
===========================

Pattern: Dependency Injection (PATTERNS_ADOPTION.md #2)
- Singleton AuthService instance
- get_current_user for protected endpoints
- Type-safe dependency injection

Success Rate: 99%
"""

from typing import Optional
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from sqlalchemy.ext.asyncio import AsyncSession

from app.database import get_db
from app.services.auth_service import AuthService
from app.models.user import User
from app.utils.security import decode_token
from jose import JWTError


# =========================================
# HTTP Bearer Security Scheme
# =========================================

security = HTTPBearer()


# =========================================
# Dependency: Get AuthService (Singleton Pattern)
# =========================================

# Singleton instance (Pattern: PATTERNS_ADOPTION.md #2)
_auth_service_instance: Optional[AuthService] = None


async def get_auth_service(
    db: AsyncSession = Depends(get_db)
) -> AuthService:
    """
    Get or create AuthService singleton
    
    Pattern: Dependency Injection (PATTERNS_ADOPTION.md #2)
    
    Benefits:
    - One instance per app lifetime (per request creates new, but service is lightweight)
    - Easy testing (can override in tests)
    - Clean code (no manual instantiation)
    
    Note: AuthService is lightweight, so we create per-request.
          For heavy services (with connections), use true singletons.
    
    Args:
        db: Database session (injected by FastAPI)
        
    Returns:
        AuthService instance
    """
    return AuthService(db)


# =========================================
# Dependency: Get Current User (Protected Routes)
# =========================================

async def get_current_user(
    credentials: HTTPAuthorizationCredentials = Depends(security),
    auth_service: AuthService = Depends(get_auth_service)
) -> User:
    """
    Get current authenticated user from JWT token
    
    Security checks (SECURITY_CHECKLIST.md):
    1. Token format valid
    2. Token not expired
    3. Token signature valid
    4. User exists in database
    5. User account is active
    6. Token not blacklisted (Phase 1: Redis)
    
    Args:
        credentials: HTTP Bearer token (injected by FastAPI)
        auth_service: AuthService instance (injected)
        
    Returns:
        User object
        
    Raises:
        HTTPException: 401 if token invalid or user not found
        
    Usage:
        @router.get("/protected")
        async def protected_endpoint(
            current_user: User = Depends(get_current_user)
        ):
            return {"user_id": current_user.id}
    """
    token = credentials.credentials
    
    try:
        # 1. Decode and validate token
        payload = decode_token(token)
        
        # 2. Extract user ID and token type
        user_id: Optional[str] = payload.get("sub")
        token_type: Optional[str] = payload.get("type")
        jti: Optional[str] = payload.get("jti")
        
        if user_id is None or token_type != "access":
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Invalid token",
                headers={"WWW-Authenticate": "Bearer"},
            )
        
        # TODO Phase 1: Check if token is blacklisted (logged out)
        # if await is_token_blacklisted(jti):
        #     raise HTTPException(
        #         status_code=status.HTTP_401_UNAUTHORIZED,
        #         detail="Token has been revoked",
        #         headers={"WWW-Authenticate": "Bearer"},
        #     )
        
        # 3. Get user from database
        user = await auth_service.get_user_by_id(user_id)
        
        if user is None:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="User not found",
                headers={"WWW-Authenticate": "Bearer"},
            )
        
        # 4. Check if user is active
        if not user.is_active:
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN,
                detail="Account is suspended",
            )
        
        return user
        
    except JWTError as e:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Could not validate credentials",
            headers={"WWW-Authenticate": "Bearer"},
        )


# =========================================
# Dependency: Optional Authentication (Public + Auth)
# =========================================

async def get_current_user_optional(
    credentials: Optional[HTTPAuthorizationCredentials] = Depends(HTTPBearer(auto_error=False)),
    auth_service: AuthService = Depends(get_auth_service)
) -> Optional[User]:
    """
    Get current user if authenticated, None if not
    
    Use for endpoints that work both authenticated and unauthenticated
    (e.g., public documents with enhanced features for authenticated users)
    
    Args:
        credentials: Optional HTTP Bearer token
        auth_service: AuthService instance
        
    Returns:
        User object if authenticated, None otherwise
        
    Usage:
        @router.get("/documents/{id}")
        async def get_document(
            id: str,
            current_user: Optional[User] = Depends(get_current_user_optional)
        ):
            # If current_user is None, show public view
            # If current_user exists, show enhanced view
            pass
    """
    if credentials is None:
        return None
    
    try:
        return await get_current_user(credentials, auth_service)
    except HTTPException:
        return None

